// lib/_TC/_SeD/_V/enums/BrdBlm_TC_SeD_V_eModel.ts
/** ---------------------------------------------------------------------------
 * @module [BrdBlm_TC_SeD_V]
 * @author [APG] Angeli Paolo Giusto
 * @version 0.1 APG 20231108
 * @version 0.2 APG 20231227 Modulo BrdBlm server side
 * ----------------------------------------------------------------------------
 */

/**
 * Modelli di portone sezionale a scorrimento verticale come da come da 
 * colonna ID della tabella:
 * [Tabelle configurazioni porte] -> [Modelli]
 */
export enum BrdBlm_TC_SeD_V_eModel {

  ARES = "18",
  ANTARES = "20",
  
  PERSUS = "37",
  ATHES = "39",
  
  CUPIS = "36",
  VENUS = "38",
  
  SIRIO = "10",
  IRIS = "11",
  
  ORUS = "54",
  
  SECURA = "61",
  DOCK = "22",
  CARGO = "23",

  LUX = "4",
  LUXOR = "5",
  FULL_VISION = "73",
}

/*! ---------------------------------------------------------------------------
 * @copyright Breda Sistemi industriali S.p.A., 2023 - http://bredasys.com
 * All rights reserved
 * @licence You cannot host, display, distribute or share this Work in any
 * form, both physical and digital. You cannot use this Work in any commercial
 * or non-commercial product, website or project. You cannot sell this Work
 * and you cannot mint an NFTs out of it.
 * ---------------------------------------------------------------------------
 */